Implemented search,delete and insert for treaps
Treaps are randomized data structure that provides a tree with expected
height of O(log N). Treaps is amalgamation of the words trees and heaps
because they show some properties from both these data structures. Treap
nodes contain (key, priority) pair. Treap nodes are organized in binary
search tree fashion on the basis of key value and are heap-ordered based
on priority. It is proven that if the priority value is randomly chosen
from a large set of numbers, then the tree will be balanced in average
case.